post api : permissions added

+++ posts/api/views.py
@@ -10,9 +10,10 @@ from django.contrib.auth.models import User
 
 from posts.models import Post
 from posts.api.serializers import PostSerializer
-
+from posts.api.permissions import PostWriterOrAdminOrReadOnly
 
 def PostListAV(APIView):
+	permission_classes = [PostWriterOrAdminOrReadOnly]
 	def get(self, request):
 
 		username = request.data.get('username')
@@ -59,6 +60,31 @@ def PostListAV(APIView):
 			return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
+def PostDetailAV(APIView):
+	permission_classes = [PostWriterOrAdminOrReadOnly]
+
+	def get(self,request):
+		username = request.data.get('username')
+		post_id  = request.data.get('post_id')
+		
+		try:
+			user_object = User.objects.get(username=username)
+		except:
+			return Response({"detail" : "User does not exists"}, status=status.HTTP_400_BAD_REQUEST)
+
+		try:
+			post_object = Post.objects.get(post_id=post_id)
+		except:
+			return Response({"detail" : "Post does not exist"}, status=status.HTTP_400_BAD_REQUEST)
+
+		self.check_object_permissions(request, post_object)
+
+		try:
+			serializer = PostSerializer(post_object)
+			return Response(serializer.data, status=status.HTTP_200_OK)
+		except:
+			return Response(serializer.error, status=status.HTTP_400_BAD_REQUEST)
+
 	def put(self, request):
 
 		username = request.data.get('username')
@@ -72,6 +98,8 @@ def PostListAV(APIView):
 			post_object = Post.objects.get(post_id=post_id)
 		except:
 			return Response({"detail" : "Post does not exist"}, status=status.HTTP_400_BAD_REQUEST)
+
+		self.check_object_permissions(request, post_object)
 
 		serializer = PostSerializer(post_object, data=request.data)
 
@@ -95,6 +123,8 @@ def PostListAV(APIView):
 			post_object = Post.objects.get(post_id=post_id)
 		except:
 			return Response({"detail" : "Post does not exist"}, status=status.HTTP_400_BAD_REQUEST)
+
+		self.check_object_permissions(request, post_object)
 
 		try:
 			post_object.delete()
